Move sos_run_script to functions.py
