+++ src/sos/hosts.py
@@ -23,7 +23,8 @@ from .syntax import SOS_LOGLINE
 from .targets import path, sos_targets
 from .task_engines import BackgroundProcess_TaskEngine
 from .tasks import TaskFile
-from .utils import (env, expand_size, expand_time, format_HHMMSS, short_repr, textMD5)
+from .utils import (env, expand_size, expand_time, format_HHMMSS, short_repr,
+                    textMD5)
 from .workflow_engines import BackgroundProcess_WorkflowEngine
 
 #
